06_deployment/streamlit/requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.7
attrs==22.1.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
boto3==1.24.51
botocore==1.27.51
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.10.0
fastjsonschema==2.16.1
filelock==3.7.1
gitdb==4.0.9
GitPython==3.1.27
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itemadapter==0.7.0
itemloaders==1.0.4
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.9.1
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parsel==1.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==4.9.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
Protego==0.2.1
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
PyDispatcher==2.0.5
Pygments==2.12.0
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
retrying==1.3.3
rich==12.5.1
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.0
Scrapy==2.6.2
semver==2.13.0
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.4.0
streamlit==1.12.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tldextract==3.3.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
Twisted==22.4.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
validators==0.20.0
w3lib==1.22.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
zipp==3.8.1
zope.interface==5.4.0
